+++ src/main/java/fr/insee/genesis/domain/service/editedresponse/editedprevious/EditedPreviousResponseJsonService.java
@@ -46,7 +46,10 @@ public class EditedPreviousResponseJsonService implements EditedPreviousResponse
             goToEditedPreviousToken(jsonParser);
             long savedCount = 0;
             Set<String> savedInterrogationIds = new HashSet<>();
-            jsonParser.nextToken(); //skip field name
+            if(jsonParser.nextToken() == null){ //skip field name, stop if end of file
+                log.warn("Reached end of file, found no EditedPrevious part.");
+                return;
+            }
             jsonParser.nextToken(); //skip [
             while (jsonParser.currentToken() != JsonToken.END_ARRAY) {
                 EditedPreviousResponseModel editedPreviousResponseModel = readNextEditedPrevious(
@@ -95,12 +98,12 @@ public class EditedPreviousResponseJsonService implements EditedPreviousResponse
         }
     }
 
-    private static void goToEditedPreviousToken(JsonParser jsonParser) throws IOException, GenesisException {
+    private static void goToEditedPreviousToken(JsonParser jsonParser) throws IOException{
         boolean isTokenFound = false;
         while (!isTokenFound){
             jsonParser.nextToken();
             if(jsonParser.currentToken() == null){
-                throw new GenesisException(400, "editedPrevious object not found in JSON");
+                return;
             }
             if(jsonParser.currentToken().equals(JsonToken.FIELD_NAME)
                     && jsonParser.currentName() != null