+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Added
 - Authorizations for scheduled Perret calls
 - Get questionnaires by review activated or not endpoint
+- Save all edited previous/external files for one questionnaire endpoint
+### Changed
+- No more 400 if the previous/external file has no corresponding part (warn in log instead)
 
 ## 1.8.4 [2025-09-11]
 ### Updated